--- NLP/list.csv
@@ -1,68 +1,0 @@
-title,url
-NAVER,https://www.naver.com/
-다른 듯 같은 북한 추석 풍경…씨름 최고 인기 : 네이버 뉴스,https://news.naver.com/main/read.nhn?mode=LSD&mid=shm&sid1=100&oid=057&aid=0001381865
-"""조국도 싫고 한국당도 싫어···20대, 정치 냉소주의 빠졌다"" : 네이버 뉴스",https://news.naver.com/main/ranking/read.nhn?mid=etc&sid1=111&rankingType=popular_day&oid=025&aid=0002936909&date=20190914&type=1&rankingSeq=2&rankingSectionId=100
-3만톤급 한국형 경항공모함' 성능 어느정도일까 : 네이버 뉴스,https://news.naver.com/main/ranking/read.nhn?mid=etc&sid1=111&rankingType=popular_day&oid=008&aid=0004277509&date=20190914&type=1&rankingSeq=9&rankingSectionId=100
-"애플 아이폰 공개 ""달라진 게 별로 없다"" : 네이버 뉴스",https://news.naver.com/main/read.nhn?mode=LSD&mid=shm&sid1=105&oid=014&aid=0004293413
-얼굴 합성 음란영상물·짤방 가능케한 검은 기술 : 네이버 뉴스,https://news.naver.com/main/ranking/read.nhn?mid=etc&sid1=111&rankingType=popular_day&oid=008&aid=0004277511&date=20190914&type=1&rankingSeq=2&rankingSectionId=105
-아이폰11 디자인에 “인덕션인가” “면도기인가” 혹평 : 네이버 뉴스,https://news.naver.com/main/read.nhn?mode=LSD&mid=shm&sid1=105&oid=014&aid=0004293264
-"추석연휴 피싱-스팸 급증..""문자 속 링크 누르지 마세요"" : 네이버 뉴스",https://news.naver.com/main/read.nhn?mode=LSD&mid=shm&sid1=105&oid=018&aid=0004467966
-겨울 스페인일주 :: 마드리드 & 세비야 여행.. : 네이버블로그,https://blog.naver.com/powerofguy2/221631359513
-:: 스페인 바로셀로나 여행 :: 겨울에 떠나.. : 네이버블로그,https://blog.naver.com/cjsxlawkd91/221645863401
-"스페인 여행 일정, 
-바르셀로나에서 시작하는 
-스페인 남부 여행! : 네이버 포스트a",https://m.post.naver.com/viewer/postView.nhn?volumeNo=22251066&memberNo=8591454&vType=VERTICAL
-초보를 위한 도커 안내서 - 설치하고 컨테이너 실행하기,https://subicura.com/2017/01/19/docker-guide-for-beginners-2.html
-초보를 위한 도커 안내서 - 도커란 무엇인가?,https://subicura.com/2017/01/19/docker-guide-for-beginners-1.html
-Docker 간편한 설치부터 실행까지,https://swalloow.github.io/docker-install
-How does Docker run a Linux kernel under macOS host? - Stack Overflow,https://stackoverflow.com/questions/43383276/how-does-docker-run-a-linux-kernel-under-macos-host
-맥(Mac)에서 이해하는 도커(Docker) #1 - junn.net,https://junn.net/archives/2323
-TDD(파이썬) : 테스트 잘하고 계신가요?,http://labs.brandi.co.kr/2018/06/07/kwakjs.html
-Stanford University CS231n: Convolutional Neural Networks for Visual Recognition,http://cs231n.stanford.edu/2017/
-Jaejun Yoo's Playground: 초짜 대학원생 입장에서 이해하는 Generative Adversarial Nets (1),http://jaejunyoo.blogspot.com/2017/01/generative-adversarial-nets-1.html
-[번역] Top 8 trends from ICLR 2019 - Subinium의 코딩일지,https://subinium.github.io/top-8-trends-from-ICLR2019-kr/
-"Topic Modeling, LDA 구현 · ratsgo's blog",https://ratsgo.github.io/from%20frequency%20to%20semantics/2017/07/09/lda/
-형태소 분석기 성능 비교 · ratsgo's blog,https://ratsgo.github.io/from%20frequency%20to%20semantics/2017/05/10/postag/
-2019 핵심 UX/UI 디자인 트렌드 모아보기 — 위디엑스,https://www.wedesignx.com/knowledge/uxui-design-trends-2019
-Introducing Slack Topics: A UX Case Study — Michelle Yang,https://www.michelleyang.me/work/slack-topics
-UX 디자인 케이스 스터디 - IT 인력 구인.. : 네이버블로그,https://blog.naver.com/sopoongp/221582804740
-"성공하는 UX 디자인의 7가지 법칙
-#4 : 네이버 포스트",https://m.post.naver.com/viewer/postView.nhn?volumeNo=22255387&memberNo=607398&vType=VERTICAL
-"책정보, 좋아보이는 것들의 비밀, UX 디자인 : 네이버 책",https://book.naver.com/bookdb/book_detail.nhn?bid=8592285
-The Best UX Design Tools in 2019 - UX Collective,https://uxdesign.cc/ux-design-tools-for-2019-ec031d060fa8
-soynlp/nounextractor-v1_usage.ipynb at master · lovit/soynlp,https://github.com/lovit/soynlp/blob/master/tutorials/nounextractor-v1_usage.ipynb
-"입문자에게 추천하는 HTML, CSS 첫걸음 | HEROPY",https://heropy.blog/2019/04/24/html-css-starter/
-Python virtualenv 정리 (Linux/Windows) :: 개발새발로그,https://dgkim5360.tistory.com/entry/python-virtualenv-on-linux-ubuntu-and-windows
-도커(Docker)를 이용하여 텐서플로우(Tensorflow) GPU 버전 설치하기 - Ubuntu 16.04(Titan XP) | devYurim,https://devyurim.github.io/python/tensorflow/development%20enviroment/docker/2018/05/25/tensorflow-3.html
-Boilerpipe tutorial Archives - Basics Behind,http://www.basicsbehind.com/tag/boilerpipe-tutorial/
-보니의 참쉬운 레시피 : 네이버 블로그,https://theboni.blog.me/221578334994
-일상기록 : 네이버 블로그,https://blog.naver.com/llzzinll?Redirect=Log&logNo=221563001369
-[백종원의 백종원 레시피] 강식당2 화제의 메뉴! 김치밥이 피오씁니다 - YouTube,https://www.youtube.com/watch?v=R6IT_f0XPT8
-다이어트식단♥닭고기케밥,http://www.10000recipe.com/recipe/6891286
-☆미라클 여행리포트☆ : 네이버 블로그,https://blog.naver.com/withjoy79/221460320255
-【일일투어】 치앙마이 화이트템플 투어 - KKday,https://www.kkday.com/ko/product/20850
-D+16 ⑤ 치앙마이 로컬 맛집 아룬 라이 A.. : 네이버블로그,https://blog.naver.com/sesool/221609213719
-Visual Studio Code 를 사용한 개발환경 꾸미기 :: C언어 예술가,https://thrillfighter.tistory.com/419
-[Elasticsearch] 기본 개념잡기 :: victolee,https://victorydntmd.tistory.com/308
-ELKR (ElasticSearch + Logstash + Kibana + Redis) 를 이용한 로그분석 환경 구축하기,https://medium.com/chequer/elkr-elasticsearch-logstash-kibana-redis-%EB%A5%BC-%EC%9D%B4%EC%9A%A9%ED%95%9C-%EB%A1%9C%EA%B7%B8%EB%B6%84%EC%84%9D-%ED%99%98%EA%B2%BD-%EA%B5%AC%EC%B6%95%ED%95%98%EA%B8%B0-f3dd9dfae622
-[번역]빠르게 배워보는 Node.js를 이용한 서버리스(Serverless) - Jiwon Yeom - Medium,https://medium.com/@jwyeom63/%EB%B9%A0%EB%A5%B4%EA%B2%8C-%EB%B0%B0%EC%9B%8C%EB%B3%B4%EB%8A%94-node-js%EB%A5%BC-%EC%9D%B4%EC%9A%A9%ED%95%9C-%EC%84%9C%EB%B2%84%EB%A6%AC%EC%8A%A4-serverless-503ee61539d4
-우린 Git-flow를 사용하고 있어요 - 우아한형제들 기술 블로그,http://woowabros.github.io/experience/2017/10/30/baemin-mobile-git-branch-strategy.html
-프로그래밍 언어별 딥러닝 라이브러리 정리 – Team AI Korea – We do Artificial Intelligence,https://aikorea.org/blog/dl-libraries/
-AI 개발에 가장 적합한 5가지 프로그래밍 언어 - ITWorld Korea,http://www.itworld.co.kr/news/109189
-CNN과 RNN의 차이 : 네이버 블로그,https://m.blog.naver.com/PostView.nhn?blogId=dhjoo98&logNo=221200176721&proxyReferer=https%3A%2F%2Fwww.google.com%2F
-[Keras Study] 8장. 생성 모델을위한 딥러닝 - Subinium의 코딩일지,https://subinium.github.io/Keras-8/
-"현대엠엔소프트 공식 기업 블로그 :: 인공지능(AI), 자동차와 만나다",https://blog.hyundai-mnsoft.com/1444
-2019에도 변하지 않을 UI/UX 트렌드 일곱가지 – 이모션,https://www.emotion.co.kr/2019%EC%97%90%EB%8F%84%EB%B3%80%ED%95%98%EC%A7%80%EC%95%8A%EC%9D%84uiux%ED%8A%B8%EB%A0%8C%EB%93%9C%EC%9D%BC%EA%B3%B1%EA%B0%80%EC%A7%80/
-UX 프로세스 배우기,https://brunch.co.kr/@ebprux/198
-미리 보는 UI 디자인 툴 전쟁,https://brunch.co.kr/@ultra0034/95
-UX 방법론이 중요한 게 아니에요,https://brunch.co.kr/@jidesign/38
-모바일 디자인 모범 사례,https://brunch.co.kr/@ebprux/242
-"부산토박이가 알려주는 2박3일 부산여행 | 세상의 모든 여행, 위시빈",https://www.wishbeen.co.kr/plan/947c830fb7e358f5
-"부산 1박2일 여행코스, 똑같은 여행은 이제 그만! 부산 여행지 추천 BEST 10",https://travelblog.expedia.co.kr/12781
-오동통그녀의 소심한 일상탈출 : 네이버 블로그,https://blog.naver.com/suheeryu/221522908919
-스위스 혼자 여행 경비 : 지식iN,https://kin.naver.com/qna/detail.nhn?d1id=9&dirId=9020305&docId=332099033&qb=7Iqk7JyE7IqkIOyXrO2WiQ==&enc=utf8&section=kin&rank=1&search_sort=0&spq=0&pid=USM2edprvh8ssLKQxMwssssssqs-464242&sid=wOe1Zsrw42hkIg90frgBRw%3D%3D
-소래비로에 오래된새댁 : 네이버 블로그,https://dlwndud1974.blog.me/221589190612
-"백종원 양파덮밥 만드는 법, 건강을 담은 한 .. : 네이버블로그",https://blog.naver.com/nahlee8/221561632175
-간단하게 맛있는 토마토 스파게티 황금레시피 : 네이버 포스트,https://m.post.naver.com/viewer/postView.nhn?volumeNo=20802755&memberNo=40599499&vType=VERTICAL
-"풍문으로 들었던 편의점 레시피, 마크정식 실천편! : 네이버 포스트",https://m.post.naver.com/viewer/postView.nhn?volumeNo=13530872&memberNo=15460571&vType=VERTICAL
-[JMT 다이어트 레시피] 가성비 만렙 밀프렙  : 네이버 포스트,https://m.post.naver.com/viewer/postView.nhn?volumeNo=20065576&memberNo=15617358&vType=VERTICAL
