+++ NLP/result/words.csv
@@ -1,0 +1,4 @@
+title,url,TF,total
+NAVER,https://www.naver.com/,"{'자동': 0.10416666666666667, '완성': 0.10416666666666667, '저장': 0.10416666666666667, '검색어': 0.10416666666666667, '도움': 0.052083333333333336, '기능': 0.052083333333333336, '신고': 0.041666666666666664, '설정': 0.03125, '기록': 0.03125, '삭제': 0.03125, '언론사': 0.03125, '사용': 0.020833333333333332, '전체': 0.020833333333333332, '추가': 0.020833333333333332, '동일': 0.010416666666666666, '시간': 0.010416666666666666, '연령': 0.010416666666666666, '남녀별': 0.010416666666666666, '그룹': 0.010416666666666666, '관심사': 0.010416666666666666, '제공': 0.010416666666666666, '해당': 0.010416666666666666, '기기': 0.010416666666666666, '브라우저': 0.010416666666666666, '현재': 0.010416666666666666, '활성': 0.010416666666666666, '공용': 0.010416666666666666, '개인': 0.010416666666666666, '정보': 0.010416666666666666, '보호': 0.010416666666666666, '로그아웃': 0.010416666666666666, '완료': 0.010416666666666666, '최대': 0.010416666666666666, '기존': 0.010416666666666666, '기준': 0.010416666666666666, '구독': 0.010416666666666666, '기사': 0.010416666666666666, 'connect': 0.010416666666666666}","[('자동', 0.10416666666666667), ('완성', 0.10416666666666667), ('저장', 0.10416666666666667), ('검색어', 0.10416666666666667), ('도움', 0.052083333333333336), ('기능', 0.052083333333333336), ('신고', 0.041666666666666664), ('설정', 0.03125), ('기록', 0.03125), ('삭제', 0.03125), ('언론사', 0.03125), ('사용', 0.020833333333333332), ('전체', 0.020833333333333332), ('추가', 0.020833333333333332), ('동일', 0.010416666666666666), ('시간', 0.010416666666666666), ('연령', 0.010416666666666666), ('남녀별', 0.010416666666666666), ('그룹', 0.010416666666666666), ('관심사', 0.010416666666666666), ('제공', 0.010416666666666666), ('해당', 0.010416666666666666), ('기기', 0.010416666666666666), ('브라우저', 0.010416666666666666), ('현재', 0.010416666666666666), ('활성', 0.010416666666666666), ('공용', 0.010416666666666666), ('개인', 0.010416666666666666), ('정보', 0.010416666666666666), ('보호', 0.010416666666666666), ('로그아웃', 0.010416666666666666), ('완료', 0.010416666666666666), ('최대', 0.010416666666666666), ('기존', 0.010416666666666666), ('기준', 0.010416666666666666), ('구독', 0.010416666666666666), ('기사', 0.010416666666666666), ('connect', 0.010416666666666666)]"
+다른 듯 같은 북한 추석 풍경…씨름 최고 인기 : 네이버 뉴스,https://news.naver.com/main/read.nhn?mode=LSD&mid=shm&sid1=100&oid=057&aid=0001381865,"{'언론사': 0.03418803418803419, '기사': 0.03418803418803419, '해당': 0.017094017094017096, '폰트': 0.017094017094017096, '설치': 0.017094017094017096, '섹션': 0.017094017094017096, '정보': 0.017094017094017096, '분류': 0.017094017094017096, '주요': 0.008547008547008548, '선정': 0.008547008547008548, '편집': 0.008547008547008548, '설정': 0.008547008547008548, '저장': 0.008547008547008548, '고딕': 0.008547008547008548, '개별': 0.008547008547008548, '이상': 0.008547008547008548, '중복': 0.008547008547008548, '추출': 0.008547008547008548, '일시': 0.008547008547008548, '최신': 0.008547008547008548, '제공': 0.008547008547008548, '기준': 0.008547008547008548, '상처': 0.008547008547008548, '댓글': 0.008547008547008548, '세상': 0.008547008547008548, '참여': 0.008547008547008548, '부탁': 0.008547008547008548, '콘텐츠': 0.008547008547008548, '저작': 0.008547008547008548, '공처': 0.008547008547008548, '무단': 0.008547008547008548, '이용': 0.008547008547008548, '경우': 0.008547008547008548, '저작권법': 0.008547008547008548, '책임': 0.008547008547008548, '북한': 0.08547008547008547, '추석': 0.08547008547008547, '풍경': 0.08547008547008547, '씨름': 0.08547008547008547, '최고': 0.08547008547008547, '인기': 0.08547008547008547, '뉴스': 0.08547008547008547}","[('북한', 0.08547008547008547), ('추석', 0.08547008547008547), ('풍경', 0.08547008547008547), ('씨름', 0.08547008547008547), ('최고', 0.08547008547008547), ('인기', 0.08547008547008547), ('뉴스', 0.08547008547008547), ('언론사', 0.03418803418803419), ('기사', 0.03418803418803419), ('해당', 0.017094017094017096), ('폰트', 0.017094017094017096), ('설치', 0.017094017094017096), ('섹션', 0.017094017094017096), ('정보', 0.017094017094017096), ('분류', 0.017094017094017096), ('주요', 0.008547008547008548), ('선정', 0.008547008547008548), ('편집', 0.008547008547008548), ('설정', 0.008547008547008548), ('저장', 0.008547008547008548), ('고딕', 0.008547008547008548), ('개별', 0.008547008547008548), ('이상', 0.008547008547008548), ('중복', 0.008547008547008548), ('추출', 0.008547008547008548), ('일시', 0.008547008547008548), ('최신', 0.008547008547008548), ('제공', 0.008547008547008548), ('기준', 0.008547008547008548), ('상처', 0.008547008547008548), ('댓글', 0.008547008547008548), ('세상', 0.008547008547008548), ('참여', 0.008547008547008548), ('부탁', 0.008547008547008548), ('콘텐츠', 0.008547008547008548), ('저작', 0.008547008547008548), ('공처', 0.008547008547008548), ('무단', 0.008547008547008548), ('이용', 0.008547008547008548), ('경우', 0.008547008547008548), ('저작권법', 0.008547008547008548), ('책임', 0.008547008547008548)]"
+"""조국도 싫고 한국당도 싫어···20대, 정치 냉소주의 빠졌다"" : 네이버 뉴스",https://news.naver.com/main/ranking/read.nhn?mid=etc&sid1=111&rankingType=popular_day&oid=025&aid=0002936909&date=20190914&type=1&rankingSeq=2&rankingSectionId=100,"{'언론사': 0.041237113402061855, '기사': 0.041237113402061855, '해당': 0.020618556701030927, '폰트': 0.020618556701030927, '설치': 0.020618556701030927, '섹션': 0.020618556701030927, '정보': 0.020618556701030927, '분류': 0.020618556701030927, '주요': 0.010309278350515464, '선정': 0.010309278350515464, '편집': 0.010309278350515464, '설정': 0.010309278350515464, '저장': 0.010309278350515464, '고딕': 0.010309278350515464, '개별': 0.010309278350515464, '이상': 0.010309278350515464, '중복': 0.010309278350515464, '추출': 0.010309278350515464, '일시': 0.010309278350515464, '최신': 0.010309278350515464, '제공': 0.010309278350515464, '기준': 0.010309278350515464, '상처': 0.010309278350515464, '댓글': 0.010309278350515464, '세상': 0.010309278350515464, '참여': 0.010309278350515464, '부탁': 0.010309278350515464, '콘텐츠': 0.010309278350515464, '저작': 0.010309278350515464, '공처': 0.010309278350515464, '무단': 0.010309278350515464, '이용': 0.010309278350515464, '경우': 0.010309278350515464, '저작권법': 0.010309278350515464, '책임': 0.010309278350515464, '조국': 0.10309278350515463, '한국': 0.10309278350515463, '정치': 0.10309278350515463, '냉소주의': 0.10309278350515463, '뉴스': 0.10309278350515463}","[('조국', 0.10309278350515463), ('한국', 0.10309278350515463), ('정치', 0.10309278350515463), ('냉소주의', 0.10309278350515463), ('뉴스', 0.10309278350515463), ('언론사', 0.041237113402061855), ('기사', 0.041237113402061855), ('해당', 0.020618556701030927), ('폰트', 0.020618556701030927), ('설치', 0.020618556701030927), ('섹션', 0.020618556701030927), ('정보', 0.020618556701030927), ('분류', 0.020618556701030927), ('주요', 0.010309278350515464), ('선정', 0.010309278350515464), ('편집', 0.010309278350515464), ('설정', 0.010309278350515464), ('저장', 0.010309278350515464), ('고딕', 0.010309278350515464), ('개별', 0.010309278350515464), ('이상', 0.010309278350515464), ('중복', 0.010309278350515464), ('추출', 0.010309278350515464), ('일시', 0.010309278350515464), ('최신', 0.010309278350515464), ('제공', 0.010309278350515464), ('기준', 0.010309278350515464), ('상처', 0.010309278350515464), ('댓글', 0.010309278350515464), ('세상', 0.010309278350515464), ('참여', 0.010309278350515464), ('부탁', 0.010309278350515464), ('콘텐츠', 0.010309278350515464), ('저작', 0.010309278350515464), ('공처', 0.010309278350515464), ('무단', 0.010309278350515464), ('이용', 0.010309278350515464), ('경우', 0.010309278350515464), ('저작권법', 0.010309278350515464), ('책임', 0.010309278350515464)]"
